Add a background story for the game
The story was generated by Chat GPT 3 based on the following prompt.

Write a story about a warrior named Danto. His small town of Cormin is
being threatened by a growing population of giant rats. The giant rats
are secretly being controlled by an evil druid named Jelakor. One day,
Danto is summoned by the village elder, Jerlian. Jerlian gives Danto a
mission to find the cause of the giant rat infestation and to bring it
to an end.